use size= option as scale factor when size_column is used
git-svn-id: https://svn.osgeo.org/grass/grass/trunk@41252 15284696-431f-4ddb-bdfa-cd5b030d7da7

+++ display/d.vect/local_proto.h
@@ -5,7 +5,7 @@ FILE *open_vect(char *, char *);
 int close_vect(FILE *);
 int plot1(struct Map_info *, int, int, struct cat_list *,
 	  const struct color_rgb *, const struct color_rgb *, int, char *,
-	  int, char *, char *, int, int, int, char *, int, char *, double,
+	  double, char *, char *, int, int, int, char *, int, char *, double,
 	  int, char *);
 int label(struct Map_info *, int, int, struct cat_list *, LATTR *, int);
 int topo(struct Map_info *, int, int, LATTR *);
